decomp: NuSound async decode path — decode thread, OGG ov_read loop, memory manager, host OpenSL fake device (audio test PASS)
Decompilation (from libTTapp.so):
- NuSoundMemoryManager: real best-fit pool allocator (Alloc/SplitFreeBuffer/
  Free/MergeFreeBuffer/CheckAndMerge*), NuSoundMemoryBuffer header contract
  (prev/next/flags bit6 alloced, bit7 locked; address field = data region),
  PopFreeBuffer/PushFreeBuffer over SCRATCH; GetSize/GetUsed/Release
  transcribed; debug helpers (OutputList/OutputMap/RenderMap/CheckList/
  AllocAddress/FreeAddress/Swap*/defrag helpers) kept as documented stubs.
- NuSoundDecodeThread: 128-slot loader ring with producer tail_index /
  consumer head_index, ThreadFunc (0x31f190) keeping the buffer's data lock
  held after SubmitBuffer like the original, request completion via the
  decoder's decode-sync pair (mutex +0xdc / cond +0xe0 / done +0xe4,
  0x31f1c8) - never the voice-held lifetime mutex.
- NuSoundDecoder::Initialise (0x31f0e0) wired into NuSound3Init after
  NuSoundSystem::Initialise, as at 0x3109aa-0x3109af; Initialise is static.
- NuSoundDecoderOGG::Decode/DecodeOggChunk: ov_read loop with the original's
  negative-returns-continue (0x32e598) and loop restart semantics.
- NuSoundLoaderOGG::ReadHeader: ov_open_callbacks with the full callback set
  (read/seek/tell/close, as the original's four callback globals); callbacks
  typed to match vorbisfile's signatures.
- NuSoundVoice/NuVoiceAndroid: loop bit lives in 0x30 bit3 (flags2 in-tree),
  SubmitBuffer sets 0x30 bit2 on short buffers and 0x17e bit1 on empty ones
  (0x32bd0e/0x32bd16), OnPlayerEvent writes hardware_flags (+0x17e).
- ReleaseVoice runs the voice's complete destructor (no free) before
  FreeMemory(SCRATCH), matching 0x31b394.
- NuSoundDecoder::GetMaxBufferSize is non-const on device; NuSoundDecoderOGG
  does not override GetName (vtable slot points at the base weak impl); the
  stub class copy in nu2api_nusound_types.h is gone (real one lives in
  nusound_decoder_ogg.hpp).

Host test rig (device boundary only, no hacks inside decompiled code):
- src/host-tests/nusound/opensl_host.cpp: SDL3-backed OpenSL fake device -
  slot-faithful vtables dispatched by offset, per-buffer-consumed events and
  HEADATEND on underrun, 44.1k->48k conversion; deleted the malloc stub
  memory manager that broke the header/address contract.
- test_audio: PASS - 2 players, 637 kB consumed (>= 576000), rms 5209 (> 30).

Gates: gonk split clean; objdiff NuMain 48.369% (>= 48.23%); check_symbols
missing 0, extras 147 (baseline bumped: the OGG decode path links two more
bundled-vorbis C-linkage symbols); clang-tidy clean.

+++ src/nu2api/nusound/nusound_loader_ogg.hpp
@@ -51,9 +51,9 @@ class NuSoundLoaderOGG : public NuSoundLoader {
     bool SeekRawData(u64 position);
     void Close();
 
-    static i32 OggCallbackClose(void *callbacks);
-    static i32 OggCallbackSeek(void *callbacks, i64 offset, i32 origin);
-    static i32 OggCallbackTell(void *callbacks);
+    static int OggCallbackClose(void *callbacks);
+    static int OggCallbackSeek(void *callbacks, i64 offset, i32 origin); // NOLINT(google-runtime-int)
+    static long OggCallbackTell(void *callbacks);                        // NOLINT(google-runtime-int)
 
   private:
     OGGFileCallbacks file_callbacks;
